arch/risc-v: Merge rv32im and rv64gc into common

+++ arch/risc-v/src/rv32m1/rv32m1_start.c
@@ -60,7 +60,7 @@
  * address.
  */
 
-uint32_t g_idle_topstack = RV32M1_IDLESTACK_TOP;
+uintptr_t g_idle_topstack = RV32M1_IDLESTACK_TOP;
 
 /****************************************************************************
  * Public Functions
